refactor(kube): isolate client-java wiring behind @ConditionalOnClass (#776)
Split JhelmKubeAutoConfiguration into a client-agnostic parent and a
ClientJavaKubeConfiguration gated on @ConditionalOnClass(ApiClient), so a
second Kubernetes backend (Fabric8) can be added and selected by whichever
client library is on the consumer's classpath.

- ClientJavaKubeConfiguration: all io.kubernetes.* beans (KubeClient,
  KubernetesProvider, KubeService, health indicator) + kubeconfig/context
  loading, active only when the official client-java is present.
- KubeServiceDecorators: extracts the client-agnostic retry/metrics decorator
  chain so every backend wraps its base service identically.
- JhelmKubeAutoConfiguration: annotations-only parent (properties binding +
  @Import of the backend config); no io.kubernetes.* reference at class level.

Pure refactor, no behavior change: existing tests green, verified against
jhelm-cli (no actuator). Part of #775.
